+++ R/smote_impl.R
@@ -86,8 +86,9 @@ smote_impl <- function(df, var, k, over_ratio) {
 
     synthetic <- smote_data(minority, k = k, n_samples = samples_needed[i])
     out_df <- as.data.frame(synthetic)
+    names(out_df) <- setdiff(names(df), var)
     out_df[var] <- data[[names(samples_needed)[i]]][[var]][1]
-    names(out_df) <- names(df)
+    out_df <- out_df[names(df)]
     out_dfs[[i]] <- out_df
   }